Update TenantController and RamlsController tests, fixing requests as needed.
Make `unsupportedAccept()` a variable argument function to allow for multiple content types in the `Accept` header.

Have the `TenantController` provide the `Accept` header validation.
- This behavior is setup to match the actual returns results (which some are in `text/plain` and others are in `application/json`).

The `spring-tenant` module now utilizes the `spring-test` module.

+++ tenant/src/main/java/org/folio/spring/tenant/controller/TenantController.java
@@ -1,20 +1,22 @@
 package org.folio.spring.tenant.controller;
 
+import static org.folio.spring.web.utility.RequestHeaderUtility.unsupportedAccept;
 import static org.springframework.http.HttpStatus.CREATED;
 import static org.springframework.http.HttpStatus.NO_CONTENT;
 
 import java.io.IOException;
 import java.sql.SQLException;
-
 import org.folio.spring.tenant.annotation.TenantHeader;
 import org.folio.spring.tenant.hibernate.HibernateSchemaService;
 import org.folio.spring.tenant.model.request.TenantAttributes;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -22,26 +24,34 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/_/tenant")
 public class TenantController {
 
+  private static final MediaType TEXT_PLAIN = new MediaType("text", "plain");
+
   @Autowired
   private HibernateSchemaService hibernateSchemaService;
 
   @PostMapping
-  // @formatter:off
   public ResponseEntity<String> create(
     @TenantHeader String tenant,
-    @RequestBody @Validated TenantAttributes attributes
+    @RequestBody @Validated TenantAttributes attributes,
+    @RequestHeader(value = "accept", required = false) String accept
   ) throws SQLException, IOException {
-  // @formatter:on
+    if (accept != null && unsupportedAccept(accept, TEXT_PLAIN)) {
+      return ResponseEntity.status(415).build();
+    }
+
     hibernateSchemaService.createTenant(tenant);
     return new ResponseEntity<String>("Success", CREATED);
   }
 
   @DeleteMapping
-  // @formatter:off
   public ResponseEntity<Void> delete(
-      @TenantHeader String tenant
+    @TenantHeader String tenant,
+    @RequestHeader(value = "accept", required = false) String accept
   ) throws SQLException {
-  // @formatter:on
+    if (accept != null && unsupportedAccept(accept, TEXT_PLAIN)) {
+      return ResponseEntity.status(415).build();
+    }
+
     hibernateSchemaService.deleteTenant(tenant);
     return new ResponseEntity<Void>(NO_CONTENT);
   }
